backend/requirements.txt
alembic==1.16.1
annotated-types==0.7.0
anyio==4.9.0
asyncpg==0.30.0
click==8.2.1
fastapi==0.115.12
greenlet==3.2.2
h11==0.16.0
idna==3.10
Mako==1.3.10
MarkupSafe==3.0.2
pydantic==2.11.5
pydantic_core==2.33.2
python-dotenv==1.1.0
sniffio==1.3.1
SQLAlchemy==2.0.41
starlette==0.46.2
typing-inspection==0.4.1
typing_extensions==4.13.2
uvicorn[standard]==0.24.0
passlib==1.7.4
PyJWT==2.10.1
python-multipart==0.0.20
